requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
lxml==4.9.3
psycopg2-binary==2.9.7
requests==2.31.0
soupsieve==2.5
tqdm==4.66.1
urllib3==2.0.4
